refactor: 수식 계산 로직 Calculator로 위임
