+++ src/main/java/Application.java
@@ -1,7 +1,8 @@
 import io.Console;
+import model.Calculator;
 import model.Expression;
+import repository.ExpressionRepository;
 import repository.MemoryExpressionRepository;
-import service.ExpressionService;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -9,19 +10,22 @@ import java.util.stream.Collectors;
 public class Application {
     public static void main(String[] args) {
         Console console = new Console();
-        ExpressionService expressionService = new ExpressionService(new MemoryExpressionRepository());
+        ExpressionRepository expressionRepository = new MemoryExpressionRepository();
 
         while(true) {
             int choice = console.choiceInput();
             if(choice == 1) {
-                List<Expression> expressions = expressionService.findAllExpression();
+                List<Expression> expressions = expressionRepository.findAll();
                 console.printAllExpressions(expressions
                         .stream()
                         .map(exp -> exp.getExpression() + " = " + exp.getCalcResult())
                         .collect(Collectors.toList()));
             } else if(choice == 2) {
                 String exprInput = console.expressionInput();
-                double calculatedNum = expressionService.calculateExpression(exprInput);
+                Expression expression = new Expression(exprInput);
+                String[] splitExpressions = exprInput.split(" ");
+                double calculatedNum = Calculator.calculate(exprInput);
+                expressionRepository.save(expression);
                 console.printCalculatedNumber(calculatedNum);
             } else {
                 console.printChooseWrongNumber();
